% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesTrendRenewal-package.R
\docType{package}
\name{BayesTrendRenewal-package}
\alias{BayesTrendRenewal}
\alias{BayesTrendRenewal-package}
\title{BayesTrendRenewal: Baysian Analysis of Renewal and Trend Renewal Processes}
\description{
Fit hierarchical renewal and trend renewal processes to the eruption records of multiple volcanoes using JAGS. Find statistical analogues for a target volcano and use these to help produce forecasts for the next eruption at the target volcano. Use model averaging to combine the posterior distribution for the forecast times produced by different models.
}
\author{
\strong{Maintainer}: Ting Wang \email{ting.wang@otago.ac.nz}

Authors:
\itemize{
  \item Joel Carman \email{joeldancarman@gmail.com}
}

}
\keyword{internal}
