% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycTRP_fit.R
\name{cycTRP_fit}
\alias{cycTRP_fit}
\title{Fit cyclic TRP model to data}
\usage{
cycTRP_fit(volNames, VEI, period, chains = 3, chainLen, adapt)
}
\arguments{
\item{volNames}{Names of volcanoes to fit (target first)}

\item{VEI}{Minimum eruption size}

\item{period}{Time period, e.g. period = 1000, use data from the last 1000 years}

\item{chains}{Number of MCMC chains to run. Default = 3}

\item{chainLen}{Length of each chain}

\item{adapt}{Length of adapt phase for each chain}
}
\value{
Posterior distributions for the parameters of the Cyclic TRP model (mcmc object)
}
\description{
Uses Bayesian methods to fit a trend renewal process model with a cyclic trend function to a set of volcanoes
}
\examples{
## Fit cyclic model to Taranaki (Target) and analogue volcanoes
## Chain length may need to be longer to get convergence
volnames <- c("Taranaki", topLast1000[[1]])
cyc.mcmc <- cycTRP_fit(volnames, 3, 1000, 3, 5000, 1000)
}
