% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_Ave.R
\name{forecast_Ave}
\alias{forecast_Ave}
\title{Model averaged forecasts}
\usage{
forecast_Ave(forecasts, weights)
}
\arguments{
\item{forecasts}{A Matrix. Each column is the posterior distribution of the forecast time from a different model}

\item{weights}{Model weights. \code{length(weights)} must be equal to \code{ncol(forecasts)}}
}
\value{
Model averaged posterior distribution for the forecast time
}
\description{
Get posterior distribution for the model averaged forecasts
}
\examples{
mod1 <- rgamma(100, 2, 2)
mod2 <- rgamma(100, 4, 4)
mod3 <- rweibull(100, 2, 1)
modAve <- forecast_Ave(cbind(mod1, mod2, mod3), weights = c(0.6, 0.3, 0.1))
}
