% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_cyc.R
\name{forecast_cyc}
\alias{forecast_cyc}
\title{Forecast using Cyclic TRP model}
\usage{
forecast_cyc(out, censorTime, firsterp, lasterp, Vol)
}
\arguments{
\item{out}{MCMC output for cyclic model}

\item{censorTime}{Censor time for target volcano (years since first eruption)}

\item{firsterp}{First eruption time for target volcano (years)}

\item{lasterp}{Last eruption time for target volcano (years since first eruption)}

\item{Vol}{Number of volcanoes used in model fitting (target + analogues)}
}
\value{
Posterior distribution for the forecast time (years)
}
\description{
Calculates the posterior distribution for the forecast time for the next eruption at the target volcano using the cyclic TRP model
}
\examples{
## Get median and 95 \% credible interval for forecast time at Taranaki using analogue volcanoes
## Chain length may need to be longer to get convergence
dat <- readGVP("Taranaki", 3, 1000) # Read taranaki data
volnames <- c("Taranaki", topLast1000[[1]])
cyc.mcmc <- cycTRP_fit(volnames, 3, 1000, 3, 5000, 1000)
cyc.fore <- forecast_cyc(cyc.mcmc, dat$cens, dat$firsterp, dat$data[dat$erp,2], length(volnames))
quantile(cyc.fore, c(0.025, 0.5, 0.975))
}
