% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_RP.R
\name{forecast_RP}
\alias{forecast_RP}
\title{Forecast using Gamma or Weibull RP models}
\usage{
forecast_RP(mcmcOut, nVol, lasterp)
}
\arguments{
\item{mcmcOut}{MCMC output for Gamma or Weibull model}

\item{nVol}{Number of volcanoes used in model fitting (target + analogues)}

\item{lasterp}{last recorded eruption time at target volcano (years)}
}
\value{
Posterior distribution for the forecast time (years)
}
\description{
Calculates the posterior distribution for the forecast time for the next eruption at the target volcano using the Gamma or Weibull RP model
}
\examples{
## Get median and 95 \% credible interval for forecast time at Taranaki using analogue volcanoes
## Chain length may need to be longer to get convergence
dat <- readGVP("Taranaki", 3, 1000) # Read taranaki data
volnames <- c("Taranaki", topLast1000[[1]])
gam.mcmc <- gamRP_fit(volnames, 3, 1000, 3, 5000, 1000)
gam.fore <- forecast_RP(gam.mcmc, length(volnames), dat$data[dat$erp,2] + dat$firsterp)
quantile(gam.fore, c(0.025, 0.5, 0.975))
}
