% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOvlp.R
\name{getOvlp}
\alias{getOvlp}
\title{Calculate overlap between posterior distributions}
\usage{
getOvlp(x, names, cut_off)
}
\arguments{
\item{x}{A matrix with n columns. Columns 2:n are compared to column 1}

\item{names}{A vector of n names, with the first name being the target volcano}

\item{cut_off}{A value between 0 and 1. Represents the minimum overlap coefficient required for a candidate analogue to be a statistical analogue}
}
\value{
A list. First element is the overlap coefficients between columns 1 and 2:n. Second element is the names of analogues which have an overlap coefficient greater than \code{cut_off}
}
\description{
Compares the posterior distributions of analogue volcanoes to the target volcano by calculating overlap
}
\examples{
## Get names of analogue volcanoes with A_i overlap > 0.2
## Chain length may need to be longer to get convergence
volnames <- c("Taranaki", topLast10000[[1]])
gam.mcmc <- gamRP_fit(volnames, 3, 10000, 3, 5000, 1000)
Amatrix <- as.matrix(gam.mcmc[,1:length(volnames)])
ovlp <- getOvlp(Amatrix, volnames, 0.2)
}
