% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GVP_Full}
\alias{GVP_Full}
\title{GVP data}
\format{
An object of class \code{data.frame} with 11222 rows and 24 columns.
}
\usage{
GVP_Full
}
\description{
Data on all eruptions from all volcanoes in the GVP catalog
}
\references{
Global Volcanism Program, 2022. [Database] Volcanoes of the World (v. 5.0.0; 1 Nov 2022).
Distributed by Smithsonian Institution, compiled by Venzke, E. \url{https://doi.org/10.5479/si.GVP.VOTW5-2022.5.0}
}
\keyword{datasets}
