% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invPsi_cyc.R
\name{invPsi_cyc}
\alias{invPsi_cyc}
\title{Inverse of cyclic trend function}
\usage{
invPsi_cyc(mu, ph, p, o, r, c, b)
}
\arguments{
\item{mu}{mu parameter}

\item{ph}{phi parameter}

\item{p}{kappa parameter}

\item{o}{omega parameter}

\item{r}{rho parameter}

\item{c}{Censored time (time since first eruption)}

\item{b}{Transformed occurrence time (time since transformed first eruption)}
}
\value{
Occurrence time t (time since first eruption)
}
\description{
Uses the cyclic trend function to calculate an occurrence time given a transformed occurrence time
}
\examples{
dat <- readGVP("Taranaki", 3, 1000)
invPsi_cyc(0.5, 0.5, 4, 3, 3, dat$cens / 1000, 0.5)
}
