% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invPsi_mono.R
\name{invPsi_mono}
\alias{invPsi_mono}
\title{Inverse of monotonic trend function}
\usage{
invPsi_mono(mu, p, c, b)
}
\arguments{
\item{mu}{mu parameter}

\item{p}{kappa parameter}

\item{c}{Censored time (time since first eruption)}

\item{b}{Transformed occurrence time (time since transformed first eruption)}
}
\value{
Occurrence time t (time since first eruption)
}
\description{
Uses the monotonic trend function to calculate an occurrence time given a transformed occurrence time
}
\examples{
dat <- readGVP("Taranaki", 3, 1000)
invPsi_mono(0.5, 4, dat$cens / 1000, 0.5)
}
