% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelWt.R
\name{modelWt}
\alias{modelWt}
\title{Model Weight}
\usage{
modelWt(waic)
}
\arguments{
\item{waic}{WAIC values}
}
\value{
Model weights
}
\description{
Calculates model weights using WAIC values
}
\examples{
  waic <- c(-45, -40, -41, -48)
  wt <- modelWt(waic)
}
