% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoTRP_fit.R
\name{monoTRP_fit}
\alias{monoTRP_fit}
\title{Fit monotonic TRP model to data}
\usage{
monoTRP_fit(volNames, VEI, period, chains = 3, chainLen, adapt)
}
\arguments{
\item{volNames}{Names of volcanoes to fit (target first)}

\item{VEI}{Minimum eruption size}

\item{period}{Time period, e.g. period = 1000, use data from the last 1000 years}

\item{chains}{Number of MCMC chains to run}

\item{chainLen}{Length of each chain}

\item{adapt}{Length of adapt phase for each chain}
}
\value{
Posterior distributions for the parameters of the monotonic TRP model (mcmc object)
}
\description{
Uses Bayesian methods to fit a trend renewal process model with a monotonic trend function to a set of volcanoes
}
\examples{
## Fit monotonic model to Taranaki (Target) and analogue volcanoes
## Chain length may need to be longer to get convergence
volnames <- c("Taranaki", topLast1000[[1]])
mono.mcmc <- monoTRP_fit(volnames, 3, 1000, 3, 5000, 1000)
}
