% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Psi_cyc.R
\name{Psi_cyc}
\alias{Psi_cyc}
\title{Cyclic trend function}
\usage{
Psi_cyc(t, mu, phi, kappa, omega, rho, tt)
}
\arguments{
\item{t}{Occurrence time (time since first eruption)}

\item{mu}{mu parameter}

\item{phi}{phi parameter}

\item{kappa}{kappa parameter}

\item{omega}{omega parameter}

\item{rho}{rho parameter}

\item{tt}{Censored time (time since first eruption)}
}
\value{
Transformed occurrence time (time since transformed first eruption)
}
\description{
Function used to transform occurrence times using the cyclic trend function
}
\examples{
dat <- readGVP("Taranaki", 3, 1000)
Psi_cyc(dat$data[,2] / 1000, mu = 0.5, phi = 0.5, kappa = 4, omega = 3, rho = 3 ,dat$cens / 1000)
}
