% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Psi_mono.R
\name{Psi_mono}
\alias{Psi_mono}
\title{Monotonic trend function}
\usage{
Psi_mono(t, mu, kappa, tt)
}
\arguments{
\item{t}{Occurrence time (time since first eruption)}

\item{mu}{mu parameter}

\item{kappa}{kappa parameter}

\item{tt}{Censored time (time since first eruption)}
}
\value{
Transformed occurrence time (time since transformed first eruption)
}
\description{
Function used to transform occurrence times using the monotonic trend function
}
\examples{
dat <- readGVP("Taranaki", 3, 1000)
Psi_mono(dat$data[,2] / 1000, mu = 0.5, kappa = 4, dat$cens / 1000)
}
