% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{topLast10000}
\alias{topLast10000}
\title{Top analogues last 10,000 years}
\format{
An object of class \code{list} of length 7.
}
\usage{
topLast10000
}
\description{
Top analogue volcanoes for Mt Taranaki identified using seven different weighting
schemes with more than 10 eruptions in the last 10,000 years
}
\keyword{datasets}
