% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic_cyc.R
\name{waic_cyc}
\alias{waic_cyc}
\title{WAIC Cyclic}
\usage{
waic_cyc(out, volnames, minVEI, period, target = F)
}
\arguments{
\item{out}{Output of Cyclic model fit}

\item{volnames}{Names of volcanoes used for model fitting (target first)}

\item{minVEI}{Minimum eruption size}

\item{period}{Time period, e.g. period = 1000, use data from the last 1000 years}

\item{target}{If \code{TRUE}, use only the target volcano data to get WAIC. If \code{FALSE} use Analogues also (Default \code{FALSE})}
}
\value{
WAIC value
}
\description{
Calculates the WAIC value for the Cyclic TRP model
}
\examples{
volnames <- c("Taranaki", topLast1000[[1]])
cyc.mcmc <- cycTRP_fit(volnames, 3, 1000, 3, 5000, 1000)
cyc.waic.analog <- waic_cyc(cyc.mcmc, volnames, 3, 1000)
cyc.waic.target <- waic_cyc(cyc.mcmc, volnames, 3, 1000, target = TRUE)
}
