% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic_mono.R
\name{waic_mono}
\alias{waic_mono}
\title{WAIC Monotonic}
\usage{
waic_mono(out, volnames, minVEI, period, target = F)
}
\arguments{
\item{out}{Output of Monotonic model fit}

\item{volnames}{Names of volcanoes used for model fitting (target first)}

\item{minVEI}{Minimum eruption size}

\item{period}{Time period, e.g. period = 1000, use data from the last 1000 years}

\item{target}{If \code{TRUE}, use only the target volcano data to get WAIC. If \code{False} use Analogues also (Default \code{FALSE})}
}
\value{
WAIC value
}
\description{
Calculates the WAIC value for the Monotonic TRP model
}
\examples{
volnames <- c("Taranaki", topLast1000[[1]])
mono.mcmc <- monoTRP_fit(volnames, 3, 1000, 3, 5000, 1000)
mono.waic.analog <- waic_mono(mono.mcmc, volnames, 3, 1000)
mono.waic.target <- waic_mono(mono.mcmc, volnames, 3, 1000, target = TRUE)
}
