% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic_weib.R
\name{waic_weib}
\alias{waic_weib}
\title{WAIC Weibull}
\usage{
waic_weib(out, volnames, minVEI, period, target = F)
}
\arguments{
\item{out}{Output of Weibull model fit}

\item{volnames}{Names of volcanoes used for model fitting (target first)}

\item{minVEI}{Minimum eruption size}

\item{period}{Time period, e.g. period = 1000, use data from the last 1000 years}

\item{target}{If \code{TRUE}, use only the target volcano data to get WAIC. If False use Analogues also}
}
\value{
WAIC value
}
\description{
Calculates the WAIC value for the Weibull model
}
\examples{
volnames <- c("Taranaki", topLast1000[[1]])
weib.mcmc <- weibRP_fit(volnames, 3, 1000, 3, 5000, 1000)
weib.waic.analog <- waic_weib(weib.mcmc, volnames, 3, 1000)
weib.waic.target <- waic_weib(weib.mcmc, volnames, 3, 1000, target = TRUE)
}
